Deseasonify: move exceptions module under utils
