+++ bot/utils/exceptions.py
